Move `lookup_path` and similar into `clippy_utils::paths`
